requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
google-generativeai
streamlit==1.28.1
plotly==5.17.0
schedule==1.2.0
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2 